Add model command, cleanup api
Add a new model api and cli command which will
display the make, model, and serial number of a
node.

Add an option to showrack to display the make, model,
and serial number of each parent node.

Add caching to client objects in the convenience api
and simplify calls by using args, kwargs